update: autentikasi

+++ routes/web.php
@@ -26,48 +26,51 @@ use Illuminate\Support\Facades\Route;
 // });
 Route::get('/role-regist',[UserController::class, 'role'])->name('role');
 Route::get('/',[UserController::class,'login'])->name('login')->middleware('guest');
+Route::post('/autentikasi',[UserController::class,'autenticate'])->name('autenticate')->middleware('guest');
+Route::post('/logout',[UserController::class,'Logout'])->name('logout')->middleware('auth');
+Route::get('/home',[UserController::class,'Home'])->name('home')->middleware('auth');
 // Route::get('/hello', function () {
 //     return view('blank');
 // });
 
 
 //kelola pengunjung
-Route::get('/pengunjung',[PengunjungController::class,'listPengunjung'])->name('pengunjung.index');
-Route::get('/buku-pengunjung',[PengunjungController::class,'addPengunjung'])->name('pengunjung.create');
-Route::post('/tambah-pengunjung',[PengunjungController::class,'storePengunjung'])->name('pengunjung.store');
+Route::get('/pengunjung',[PengunjungController::class,'listPengunjung'])->name('pengunjung.index')->middleware('auth');
+Route::get('/buku-pengunjung',[PengunjungController::class,'addPengunjung'])->name('pengunjung.create')->middleware('auth');
+Route::post('/tambah-pengunjung',[PengunjungController::class,'storePengunjung'])->name('pengunjung.store')->middleware('auth');
 
 //kelola rak
-Route::get('/rak-buku',[RakKategoriController::class,'listRak'])->name('rak.index');
-Route::post('/tambah-rak',[RakKategoriController::class,'addRak'])->name('rak.store');
-Route::put('/rak-update/{id}',[RakKategoriController::class,'updateRak'])->name('rak.update');
-Route::delete('/rak-delete/{id}',[RakKategoriController::class,'deleteRak'])->name('rak.destroy');
+Route::get('/rak-buku',[RakKategoriController::class,'listRak'])->name('rak.index')->middleware('auth');;
+Route::post('/tambah-rak',[RakKategoriController::class,'addRak'])->name('rak.store')->middleware('auth');;
+Route::put('/rak-update/{id}',[RakKategoriController::class,'updateRak'])->name('rak.update')->middleware('auth');;
+Route::delete('/rak-delete/{id}',[RakKategoriController::class,'deleteRak'])->name('rak.destroy')->middleware('auth');;
 
 
 //kelola kategori
-Route::get('/kategori-buku',[RakKategoriController::class,'listKategori'])->name('kategori.index');
-Route::post('/tambah-kategori',[RakKategoriController::class,'addKategori'])->name('kategori.store');
-Route::put('/kategori-update/{id}',[RakKategoriController::class,'updateKategori'])->name('kategori.update');
-Route::delete('/kategori-delete/{id}',[RakKategoriController::class,'deleteKategori'])->name('kategori.destroy');
+Route::get('/kategori-buku',[RakKategoriController::class,'listKategori'])->name('kategori.index')->middleware('auth');;
+Route::post('/tambah-kategori',[RakKategoriController::class,'addKategori'])->name('kategori.store')->middleware('auth');;
+Route::put('/kategori-update/{id}',[RakKategoriController::class,'updateKategori'])->name('kategori.update')->middleware('auth');;
+Route::delete('/kategori-delete/{id}',[RakKategoriController::class,'deleteKategori'])->name('kategori.destroy')->middleware('auth');;
 
 //kelola buku
-Route::get('/buku', [BukuController::class,'index'])->name('buku.index');
-Route::get('/tambah_buku',[BukuController::class,'create'])->name('buku.create');
-Route::post('/tambah_buku/store',[BukuController::class,'store'])->name('buku.store');
-Route::get('/buku/info/{id}',[BukuController::class,'info'])->name('buku.info');
-Route::get('/buku/edit/{id}',[BukuController::class,'edit'])->name('buku.edit');
-Route::put('/buku/update/{id}',[BukuController::class,'update'])->name('buku.update');
-Route::delete('/buku/delete/{id}',[BukuController::class, 'destroy'])->name('buku.destroy');
+Route::get('/buku', [BukuController::class,'index'])->name('buku.index')->middleware('auth');
+Route::get('/tambah_buku',[BukuController::class,'create'])->name('buku.create')->middleware('auth');;
+Route::post('/tambah_buku/store',[BukuController::class,'store'])->name('buku.store')->middleware('auth');;
+Route::get('/buku/info/{id}',[BukuController::class,'info'])->name('buku.info')->middleware('auth');;
+Route::get('/buku/edit/{id}',[BukuController::class,'edit'])->name('buku.edit')->middleware('auth');;
+Route::put('/buku/update/{id}',[BukuController::class,'update'])->name('buku.update')->middleware('auth');;
+Route::delete('/buku/delete/{id}',[BukuController::class, 'destroy'])->name('buku.destroy')->middleware('auth');;
 
 // kelola sirkulasi
-Route::get('/peminjaman',[SirkulasiController::class,'pinjam'])->name('sirkulasi.pinjam');
-Route::get('/pengembalian',[SirkulasiController::class,'kembali'])->name('sirkulasi.kembali');
-Route::get('/tambah-pinjam',[SirkulasiController::class,'create'])->name('sirkulasi.tambah');
-Route::post('/input-pinjam',[SirkulasiController::class,'store'])->name('sirkulasi.store');
-Route::get('/pinjam/info/{id}',[SirkulasiController::class,'info'])->name('pinjam.info');
-Route::delete('/kembali/hapus/{id}',[SirkulasiController::class,'destroy'])->name('pinjam.destroy');
-Route::post('/pinjam/kembalikan/{id}',[SirkulasiController::class,'return'])->name('sirkulasi.kembalikan');
-Route::get('/pengembalian/info/{id}',[SirkulasiController::class,'detail'])->name('kembali.info');
+Route::get('/peminjaman',[SirkulasiController::class,'pinjam'])->name('sirkulasi.pinjam')->middleware('auth');;
+Route::get('/pengembalian',[SirkulasiController::class,'kembali'])->name('sirkulasi.kembali')->middleware('auth');;
+Route::get('/tambah-pinjam',[SirkulasiController::class,'create'])->name('sirkulasi.tambah')->middleware('auth');;
+Route::post('/input-pinjam',[SirkulasiController::class,'store'])->name('sirkulasi.store')->middleware('auth');;
+Route::get('/pinjam/info/{id}',[SirkulasiController::class,'info'])->name('pinjam.info')->middleware('auth');;
+Route::delete('/kembali/hapus/{id}',[SirkulasiController::class,'destroy'])->name('pinjam.destroy')->middleware('auth');;
+Route::post('/pinjam/kembalikan/{id}',[SirkulasiController::class,'return'])->name('sirkulasi.kembalikan')->middleware('auth');;
+Route::get('/pengembalian/info/{id}',[SirkulasiController::class,'detail'])->name('kembali.info')->middleware('auth');;
 
 //denda
-Route::get('/denda',[DendaController::class,'index'])->name('denda.index');
-Route::put('/denda/update/{id}',[DendaController::class,'update'])->name('denda.update');
+Route::get('/denda',[DendaController::class,'index'])->name('denda.index')->middleware('auth');;
+Route::put('/denda/update/{id}',[DendaController::class,'update'])->name('denda.update')->middleware('auth');;
